--- src/main/java/cyclone/otusspring/hw01/service/AskService.java
@@ -1,14 +1,0 @@
-package cyclone.otusspring.hw01.service;
-
-import cyclone.otusspring.hw01.model.Answer;
-import cyclone.otusspring.hw01.model.Person;
-import cyclone.otusspring.hw01.model.Question;
-
-import java.util.List;
-
-public interface AskService {
-
-    Person preparePerson();
-
-    List<Answer> askQuestions(Person person, List<Question> questions);
-}
