+++ src/main/java/cyclone/otusspring/poll/io/ConsoleUserIO.java
@@ -1,4 +1,4 @@
-package cyclone.otusspring.hw01.io;
+package cyclone.otusspring.poll.io;
 
 import org.springframework.stereotype.Component;
