+++ src/main/java/cyclone/otusspring/poll/io/UserIO.java
@@ -1,4 +1,4 @@
-package cyclone.otusspring.hw01.io;
+package cyclone.otusspring.poll.io;
 
 public interface UserIO {
     void print(String line);
