+++ src/main/java/cyclone/otusspring/poll/runner/PollRunner.java
@@ -1,4 +1,4 @@
-package cyclone.otusspring.hw01.runner;
+package cyclone.otusspring.poll.runner;
 
 public interface PollRunner {
     void run();
