+++ src/main/java/cyclone/otusspring/poll/service/ResultService.java
@@ -1,0 +1,10 @@
+package cyclone.otusspring.poll.service;
+
+import cyclone.otusspring.poll.model.Answer;
+import cyclone.otusspring.poll.model.Result;
+
+import java.util.List;
+
+public interface ResultService {
+    Result getResult(List<Answer> answers);
+}
